- rename `module.js` files to `index.js` - fix local imports to be relative
