+++ src/clip/index.js
